conda_requirements.txt
# conda requirements file for the oil_library package.

# These should all be in the conda-forge channel:
# Best to have it at the top in your conda configuration

python=2.7.*
pathlib
gitpython
numpy>=1.16.*
scipy>=0.18.1

SQLAlchemy>=1.3.7
zope.sqlalchemy>=1.1

awesome-slugify>=1.6.5
pytest>=2.9.2
backports.functools_lru_cache>=1.5
pynucos>=2.7.4

# to support py2/3 compatibility
future
